requirements.txt
beautifulsoup4==4.4.0
dateutils==0.6.6
Django==1.8.4
ftfy==4.0.0
html5lib==0.999999
jsonfield==1.0.3
lxml==3.4.4
nyt-clerk==0.0.4
psycopg2==2.6.1
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
six==1.9.0
smartypants==1.8.6
wheel==0.24.0
